move icon related classes to another package
